+++ host/libubertooth/src/ubertooth.h
@@ -75,7 +75,7 @@ typedef struct {
 void print_version();
 void register_cleanup_handler(ubertooth_t* ut);
 ubertooth_t* ubertooth_init();
-uint8_t ubertooth_connect(ubertooth_t* ut, int ubertooth_device);
+int ubertooth_connect(ubertooth_t* ut, int ubertooth_device);
 ubertooth_t* ubertooth_start(int ubertooth_device);
 void ubertooth_stop(ubertooth_t* ut);
 void ubertooth_set_timeout(ubertooth_t* ut, int seconds);
